moved util functions up to home/utils.ts
