+++ frontend/src/metabase/home/utils.ts
@@ -12,3 +12,27 @@ export const isWithinWeeks = (
   const weeksAgo = moment().subtract(weekCount, "week");
   return date.isAfter(weeksAgo);
 };
+
+export const setShowEmbedHomepageFlag = () => {
+  try {
+    localStorage.setItem("showEmbedHomepage", "true");
+  } catch (e) {
+    console.error(e);
+  }
+};
+
+export const removeShowEmbedHomepageFlag = () => {
+  try {
+    localStorage.removeItem("showEmbedHomepage");
+  } catch (e) {
+    console.error(e);
+  }
+};
+
+export const shouldShowEmbedHomepage = () => {
+  try {
+    return localStorage.getItem("showEmbedHomepage") === "true";
+  } catch (e) {
+    return false;
+  }
+};
